pyproject.toml
[project]
name = "learn-phidata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "phidata==2.7.10",
    "duckduckgo-search==8.1.1",
    "python-dotenv==1.2.2",
    "google-genai>=1.67.0",
    "google-generativeai>=0.8.6",
    "yfinance>=1.2.0",
]
